+++ src/main/java/kr1v/malilibApi/config/_new/ConfigObject.java
@@ -1,4 +1,4 @@
-package kr1v.malilibApi.config;
+package kr1v.malilibApi.config._new;
 
 import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonElement;