+++ src/testmod/java/kr1v/malilibApiTest/config/Test5.java
@@ -3,7 +3,11 @@ package kr1v.malilibApiTest.config;
 import com.google.common.collect.ImmutableList;
 import fi.dy.masa.malilib.config.options.*;
 import kr1v.malilibApi.annotation.*;
-import kr1v.malilibApi.config.*;
+import kr1v.malilibApi.config._new.ConfigList;
+import kr1v.malilibApi.config._new.ConfigPair;
+import kr1v.malilibApi.config.custom.ArrayBackedCycleConfig;
+import kr1v.malilibApi.config.custom.ConfigCycle;
+import kr1v.malilibApi.config.custom.EnumBackedCycleConfig;
 import kr1v.malilibApiTest.Init;
 import net.minecraft.world.GameMode;