+++ src/testmod/java/kr1v/malilibApiTest/custo/ConfigClass.java
@@ -1,7 +1,7 @@
 package kr1v.malilibApiTest.custo;
 
 import com.google.gson.JsonElement;
-import kr1v.malilibApi.config.CustomConfigBase;
+import kr1v.malilibApi.config._new.CustomConfigBase;
 
 public class ConfigClass extends CustomConfigBase<ConfigClass> {
     public ConfigClass(String name, String comment, String translatedName, String prettyName) {