+++ changelog.md
@@ -14,6 +14,6 @@ No new things for users
   - A config type that is a Pair of 2 `IConfigBase`s. They will show up next to each other in game, each taking up half the width of the config. Yes, they both can be another `ConfigPair`
 * Custom config option support
   - See `MalilibApi#registerWidgetBasedConfigType` and `MalilibApi#registerButtonBasedConfigType`, or these for examples:
-    - (Button based) `kr1v.malilibApi.config.ConfigButton` + `kr1v.malilibApi.widget.ConfigButtonButton`
-    - (Button based) `kr1v.malilibApi.config.ConfigList` + `kr1v.malilibApi.widget.ConfigListButton`
-    - (Widget based) `kr1v.malilibApi.config.ConfigPair` + `kr1v.malilibApi.widget.WidgetPair`
+    - (Button based) `kr1v.malilibApi.config._new.ConfigButton` + `kr1v.malilibApi.widget.ConfigButtonButton`
+    - (Button based) `kr1v.malilibApi.config._new.ConfigList` + `kr1v.malilibApi.widget.ConfigListButton`
+    - (Widget based) `kr1v.malilibApi.config._new.ConfigPair` + `kr1v.malilibApi.widget.WidgetPair`